add support for xcffib as an alternative to xpyb
xcffib is well maintained and packaged for most Linux distros, and it is
meant to be mostly a drop-in replacement to xpyb. It would be great if
xcffib support could be added so we don't need to package the legacy
xpyb.

I have also filed a pull request for xpybutil to add xcffib support at
https://github.com/BurntSushi/xpybutil/pull/8

Change-Id: I9491b5383d6df2ee214e804d93903f91484be076